+++ apps/fluent-tester/src/TestComponents/Checkbox/prev1/consts.ts
@@ -1,0 +1,12 @@
+export const HOMEPAGE_CHECKBOX_BUTTON = 'Homepage_Checkbox_Button';
+export const CHECKBOX_TESTPAGE = 'Checkbox_TestPage';
+
+/* Test Checkbox 1 */
+export const CHECKBOX_TEST_COMPONENT = 'Checkbox_Test_Component'; // A component on each specific test page
+export const CHECKBOX_ACCESSIBILITY_LABEL = 'E2E testing Checkbox accessibility label';
+
+/* Test Checkbox 2 */
+export const CHECKBOX_NO_A11Y_LABEL_COMPONENT = 'Checkbox_No_A11y_label_Component';
+export const CHECKBOX_TEST_COMPONENT_LABEL = 'Test Checkbox2 - No Accessibility Label'; // A component on each specific test page
+
+export const CHECKBOX_ON_PRESS = 'Checkbox_On_Press'; // For testing the press functionality on a Checkbox
