--- apps/fluent-tester/src/TestComponents/CheckboxExperimental/consts.ts
@@ -1,12 +1,0 @@
-export const HOMEPAGE_CHECKBOX_EXPERIMENTAL_BUTTON = 'Homepage_CheckboxExperimental_Button'; // Button on the homepage that navigates to the selected test page
-export const EXPERIMENTAL_CHECKBOX_TESTPAGE = 'CheckboxExperimental_TestPage'; // A component on each specific test page.
-
-/* Test Experimental Checkbox 1 */
-export const EXPERIMENTAL_CHECKBOX_TEST_COMPONENT = 'Experimental_Checkbox_Test_Component'; // A component on each specific test page
-export const EXPERIMENTAL_CHECKBOX_ACCESSIBILITY_LABEL = 'E2E testing Experimental Checkbox accessibility label';
-
-/* Test Experimental Checkbox 2 */
-export const EXPERIMENTAL_CHECKBOX_NO_A11Y_LABEL_COMPONENT = 'Experimental Checkbox_No_A11y_label_Component';
-export const EXPERIMENTAL_CHECKBOX_TEST_COMPONENT_LABEL = 'Test Experimental Checkbox2 - No Accessibility Label'; // A component on each specific test page
-
-export const EXPERIMENTAL_CHECKBOX_ON_PRESS = 'Experimental_Checkbox_On_Press'; // For testing the press functionality on a Checkbox
